refactor: move Riverpod providers to correct architecture layers
