+++ pubspec.yaml
@@ -282,6 +282,10 @@ dev_dependencies:
 
   riverpod_generator: 4.0.4-dev.3
 
+  custom_lint: ^0.7.5
+
+  riverpod_lint: ^3.4.0
+
   mockito: 5.6.4
 
   hive_ce_generator: 1.11.2
